pinout and tone fix
I had to come up with a workaround for the tone library, because some
library must have interfered with it. The Theremin is back on track
now! I also had to throw out the 1812 part and turn off the speaker
manualy by digitalWrite. I must have converted RCSwitch as a library or
somebody else, because I have it in my libraries directory.